SpatialOps Checkin
 - Significantimprovements to the SpatialFieldStore class.  This
should make it very easy to obtain temporary (work) fields as needed.
It also provides a mechanism to minimize the number of work fields,
and only allocates new ones on an as-deeded basis.

 - Eliminate binary operators from the SpatialField class.  These are
now only supported through the SpatFldPtr class, which can
automatically manage the temporary fields required for use in these
binary operations.



git-svn-id: https://software.crsim.utah.edu/svn/SpatialOps/trunk@118 95f76ec8-9725-0410-b239-af81322155aa