This checkin represents a major revision to SpatialOps.  Among the major changes:
 - SpatialOperator and SpatialField classes no longer have any
knowledge of the domain layout, and do not force a structured mesh.
This should allow them to be used in unstructured meshes.  There were
minor changes throughout these classes.

 - LinearSystem still has some assumptions about a structured mesh.

 - Introduced an interior_iterator for SpatialFields.  This allows one
to iterate over interior points (excluding ghost values).

 - The SpatialField constructor has changed to be more general.

 - The Staggered finite volume operators and fields have been
completely reworked.  This was required due to the huge number of
operators and field types required for a staggered FV mesh.  I still
have not tested (or coded) all of the operators.  However, a large
number (~30) have been coded and tested.  This certainly demonstrates
the advantages of generic programming!

 - The test suite needs to be cleaned up a bit.  The tests that
Alberto wrote are still there, but do not work with the new fields and
operators.  Something analogous needs to be written.

 - There is still a fair bit of work to be done to get this production
ready.  Scratch operators, and several interpolant operators have not
been coded or tested yet.


git-svn-id: https://software.crsim.utah.edu/svn/SpatialOps/trunk@150 95f76ec8-9725-0410-b239-af81322155aa