Fixed bug in Stencil2 implementation not covered in current tests Fix reflected both in s2detail::ExtentsAndOffsets and in RefStencil2Detail::ExtentsAndOffsets.
The bug comes up in a rare case where there are extra cells in both the
source and destination field normal to the same axis under a boundary
condition, and that axis is not the same axis as the active direction
of the stencil operator.

For example, consider the interpolant stencil operator from XVolField
to YSurfXField.  The active direction of this operator is the Y-axis.
However, both XVolField and YSurfField have extra boundary cells normal
to the Y-axis, when there is a boundary normal to the Y-axis.

Without the fix in this commit, the interpolant operator in the above
example does not calculate the extra cells when there is a boundary
condition.

With the fix in this commit, the extra cells from the boundary condition
are calculated.