Add NullStencil to Nebo's inlined stencil implementation
With this commit, code such as:
 NullStOp.apply_to_field(src, dest);
can be replaced with:
 dest <<= NullStOp(src);
where dest and src are SpatialFields and NullStOp is a NullStencil.

Also, chaining of stencil is allowed.
So with this commit, code such as:
 NullStOp.apply_to_field(src, tmp);
 OtherStOp.apply_to_field(tmp, dest);
can be replaced with:
 dest <<= OtherStOp(NullStOp(src));
where dest, tmp, and src are SpatialFields; NullStOp is a NullStencil, and OtherStOp is any stencil that works with Nebo's inlined stencil implementation.

This functionality has NOT been tested!
The apply_to_field of NullStencil has not been changed, so all existing code is unaffected.