Clean up code - no functional change
Rename ResizePrep, the Nebo mode, to Resize.
Essentially the two modes, ResizePrep and Resize, have been merged into one.
It was easier to fold the functionality of Resize into ResizePrep.
ResizePrep was not renamed in the same commit to avoid confusion with the merging process.

There were other minor changes in FieldExpressions.fmr which do not appear in FieldExpressions.h.
These changes were done to make FieldExpressions.fmr more consistent.