Move Nebo's CUDA backend from operator |= to operator <<=.
Operator |= is completely removed.

Several conditions must be met to for Nebo to use the CUDA backend:
 First, ENABLE_CUDA must be ON at compile-time.
 Second, the Nebo assignments (uses of <<=) must be in .cu files that are compiled by Nvidia's compiler (nvcc).
 Third, at runtime, the fields in the Nebo assignments must be allocated on the GPU.

If not all of these conditions are met, the Nebo attempts to use either the thread parallel or the sequential Nebo backends.

Also, there are minor formating changes that have no functional change.