Add NEBO_GPU_TEST mode.
NOTE: NEBO_GPU_TEST is for testing only and NOT for performance-sensitive applications.

NOTE: Nebo's GPU backend is still limited to uses of Nebo in *.cu files.

NOTE: NEBO_GPU_TEST assumes all data is on the CPU and that it handles all copying.

If the C preprocessor macro, NEBO_GPU_TEST is defined at compile-time
(as well as ENABLE_CUDA), Nebo will handle all data transfer between the CPU and GPU.

For example, consider the following Nebo statements:

 c <<= a + sin(b);
 d <<= StencilOperator(c);

If this code is in a *.cu file and is compiled with ENABLE_CUDA and NEBO_GPU_TEST, then
the first statement will copy fields, a and b, to the GPU, execute on the GPU, and copy the
result back into field, c.  The next statement will copy the field, c, back to the GPU, execute,
and copy the result back to the field, d.  Thus, there will be an unnecessary copy of c from
and then immediately back to the GPU.

NEBO_GPU_TEST is designed to test the correctness of Nebo's GPU backend without rewriting a lot of tests.