Modify NEBO_GPU_TEST to work when data is already on GPU
NEBO_GPU_TEST now can work if the assignee field is already allocated on the GPU.
If the fields on the right-hand side of the assignment already have consumer fields on
the GPU, the data is recopied, to insure the correct data is on the GPU.

For example, consider the following assignment and assume NEBO_GPU_TEST is set to ON:

d <<= a + sin(b) / c;

If d's data is allocated on the CPU, Nebo will copy the data to the GPU, calculate the
result, and copy the result back to the CPU.  If d's data is allocated on the GPU, Nebo
will only calculate the result, and not do any data transfer for d.

If a's data is on the CPU, b's is on the GPU, and c's is on the CPU with a consumer field
already on the GPU, Nebo will call add_consumer on each.  For a, add_consumer will allocate
space on the GPU and copy the data over.  For b, add_consumer will do nothing. For c,
add_consumer will only copy the data from the CPU to the GPU, using the space already
allocated.

The purpose of copying the data over to the GPU, when space is already allocated (but the
primary is on the CPU), is to use the most up-to-date data in the field.