First step toward runtime ghost specification.
This commit adds two key classes:

  1)  GhostDataRT (to be renamed to GhostData in the future).  This describes ghost cells on the (-) and (+) sides of a memory block

  2)  BoundaryCellInfo.  This holds information about whether a physical boundary is present on the (+) side of the domain and the number of extra points that are acquired.

The SpatialField constructor now requires objects of these two classes to be passed as arguments to the constructor.  These will form the basis for our runtime ghost cell treatment.

NOTE: there are many tests that fail.  This is due to nebo's current treatment of stencils, which is incompatible with this new approach.

There will be a lot of cleanup necessary to remove the compile-time ghost specification, but that will be done once we have the runtime approach fully functional.