Create SingleValueField and add support for most of Nebo
A SingleValueField is a SpatialField with two major restrictions:
 - The extent of a SingleValueField should be (1,1,1)
 - A SingleValueField cannot have any ghost cells.
There is runtime error checking for these properties in debug mode.

A SingleValueField has several advantages over a POD scalar value
and over a SpatialField of a different type with the same window:
 - A SingleValueField can have copies on CPU and GPU devices (use
   add_consumer).
 - A SingleValueField can assigned to in Nebo (on CPU or GPU).
 - A SingleValueField can be used in SpatFldPtr without template
   specialization.
 - A SingleValueField can be used on the right-hand side of any
   Nebo expression, regardless of the result type.

Currently, SingleValueField can work with all of Nebo EXCEPT cond.
A later update will add cond functionality for SingleValueField.